+++ plugin/types/versionFile.go
@@ -1,6 +1,6 @@
 package types
 
-//WriteVersionFile(v8end V8Endpoint, dir string, number int64, filename string) error
+//WriteVersionFile(v8end V8Endpoint, dir string, number int, filename string) error
 
 type WriteVersionFileSubscriber struct {
 	Before BeforeWriteVersionFileFn
@@ -8,8 +8,8 @@ type WriteVersionFileSubscriber struct {
 	After  AfterWriteVersionFileFn
 }
 type (
-	BeforeWriteVersionFileFn func(v8end V8Endpoint, workdir string, number int64, filename string) error
-	OnWriteVersionFileFn     func(v8end V8Endpoint, workdir string, number int64, filename string, stdHandler *bool) error
+	BeforeWriteVersionFileFn func(v8end V8Endpoint, workdir string, number int, filename string) error
+	OnWriteVersionFileFn     func(v8end V8Endpoint, workdir string, number int, filename string, stdHandler *bool) error
 	AfterWriteVersionFileFn  BeforeWriteVersionFileFn
 )
 
@@ -20,6 +20,6 @@ type ReadVersionFileSubscriber struct {
 }
 type (
 	BeforeReadVersionFileFn func(v8end V8Endpoint, workdir string, filename string) error
-	OnReadVersionFileFn     func(v8end V8Endpoint, workdir string, filename string, stdHandler *bool) (int64, error)
-	AfterReadVersionFileFn  func(v8end V8Endpoint, workdir string, filename string, number *int64) error
+	OnReadVersionFileFn     func(v8end V8Endpoint, workdir string, filename string, stdHandler *bool) (int, error)
+	AfterReadVersionFileFn  func(v8end V8Endpoint, workdir string, filename string, number *int) error
 )
